refactor(main): give HttpError and the Team API request helpers their own files
A pure move ahead of splitting the router. `team-api-server.ts` loses 511 lines and gains
nothing but imports; every moved line is byte-identical.

`http-error.ts` is deliberately alone in a file. The router's single catch classifies by
`instanceof`, so a second copy of the class anywhere would make that check silently false and
turn every 400 a route raises into a 500 "Request failed." It cannot live in the server either:
the route modules and the request helpers both throw it, and importing it back would put a class
declaration in an import cycle.

`dependencies.ts` takes the `Pick<>` service types. `request-helpers.ts` takes the parsers, the
validators and the capability filters - the ones a route module needs. `eventCapability` and
`unavailableSidebarLayout` stay behind with the WebSocket side that is their only caller.

The only non-move is `import("node:http").IncomingMessage` becoming a normal type import now
that the file has its own import block.

Checked: `bun run test:desktop -- src/main/team-api-server.test.ts` (39 passed),
`typecheck:node`, `biome check src/main/team-api src/main/team-api-server.ts`.
