+++ src/main/team-api-server.ts
@@ -7,12 +7,6 @@ import { ATTACHMENT_LIMITS, AVATAR_IMAGE_LIMITS, INPUT_LIMITS } from "@openbot/c
 import {
   AGENT_RUNTIME_SNAPSHOT_BYTES_LIMIT,
   type AgentEvent,
-  type CentralAuthUser,
-  type ConversationPageAnchor,
-  type ConversationSnapshot,
-  type ConversationWithReadState,
-  type CreateBotInput,
-  type CreateTeamInviteInput,
   type DirectConversationPage,
   type DirectConversationPageAnchor,
   type DirectConversationSnapshot,
@@ -21,28 +15,16 @@ import {
   type DirectThreadSummary,
   type DirectTypingRealtimeEvent,
   type DuplicateBotResult,
-  HOSTED_SITE_EVENT_ITEM_TYPE_PREFIX,
-  type InstalledSkill,
-  type InviteSummary,
-  isAgentModel,
-  isAvatarHue,
-  isAvatarSeed,
   isMessageReaction,
-  isReasoningEffort,
   type ReorderQueueInput,
-  type RespondToApprovalInput,
-  type RespondToBrowserTakeoverInput,
-  ROUTINE_EVENT_ITEM_TYPE_PREFIX,
-  ROUTINE_RUN_EVENT_ITEM_TYPE_PREFIX,
   type SidebarLayoutSnapshot,
   type SteerQueuedMessageInput,
   type TeamMemberSummary,
   type TeamPresenceSnapshot,
   type TeamRealtimeEvent,
-  type UpdateBotInput,
   type UpdateQueuedMessageInput,
 } from "@openbot/contracts/ipc";
-import { type DynamicRecord, isBoolean, isDynamicRecord, isNumber, isString } from "@openbot/contracts/runtime-values";
+import { isBoolean, isNumber, isString } from "@openbot/contracts/runtime-values";
 import { TEAM_API_ROUTES } from "@openbot/contracts/team-api-routes";
 import {
   isTeamCurrentCapability,
@@ -54,7 +36,6 @@ import {
 import {
   decodeTeamProtocolV1ClientEvent,
   TEAM_APP_VERSION_HEADER,
-  TEAM_CAPABILITIES_HEADER,
   TEAM_PROTOCOL_V1,
   TEAM_PROTOCOL_V1_CAPABILITIES,
   TEAM_PROTOCOL_V1_WEBSOCKET,
@@ -62,21 +43,13 @@ import {
   type TeamProtocolSupportV1,
 } from "@openbot/contracts/team-protocol/v1";
 import {
-  decodeTeamProtocolV1CurrentHttpRequest,
   encodeTeamProtocolV1CurrentEvent,
   encodeTeamProtocolV1CurrentHttpResponse,
 } from "@openbot/contracts/team-protocol/v1-adapter";
 import { TEAM_PROTOCOL_V3 } from "@openbot/contracts/team-protocol/v3";
-import {
-  decodeTeamProtocolV3CurrentHttpRequest,
-  encodeTeamProtocolV3CurrentHttpResponse,
-} from "@openbot/contracts/team-protocol/v3-adapter";
-import { createOpenBotLogger, type Logger, toLogValue } from "@openbot/logging";
+import { encodeTeamProtocolV3CurrentHttpResponse } from "@openbot/contracts/team-protocol/v3-adapter";
+import { createOpenBotLogger, toLogValue } from "@openbot/logging";
 import type * as Ws from "ws";
-import type { AgentService } from "../backend/agent-service";
-import type { BrowserHost } from "../backend/browser-host";
-import type { MailboxStore } from "../backend/mailbox-store";
-import type { SidebarLayoutStore } from "../backend/sidebar-layout-store";
 import type { TeamChatStore } from "../backend/team-chat-store";
 import {
   parseCreateRoutine,
@@ -84,10 +57,38 @@ import {
   parseSidebarLayoutAction,
   parseUpdateRoutine,
 } from "./ipc/agent-inputs";
-import { RemoteScreenError, type RemoteScreenGateway } from "./remote-screen-gateway";
-import { type TeamStore, TeamStoreError } from "./team-store";
+import { RemoteScreenError } from "./remote-screen-gateway";
+import type { TeamApiOptions, TeamApiSidebarLayout } from "./team-api/dependencies";
+import { HttpError } from "./team-api/http-error";
+import {
+  approvalDecision,
+  bearerToken,
+  botCreate,
+  botUpdate,
+  browserTakeoverDecision,
+  conversationForCapabilities,
+  conversationSnapshotForCapabilities,
+  firstHeaderValue,
+  JSON_LIMIT,
+  markerExclusionsForCapabilities,
+  nullableString,
+  pageAnchor,
+  pageLimit,
+  parseBrowserBounds,
+  pathIdentifier,
+  promptAnswers,
+  promptRequestId,
+  publicHttpBaseUrl,
+  readBinary,
+  readJson,
+  requestCapabilities,
+  requestProtocol,
+  requireAdmin,
+  stringArray,
+  stringField,
+} from "./team-api/request-helpers";
+import { TeamStoreError } from "./team-store";
 
-const JSON_LIMIT = 1024 * 1024;
 const EVENT_PAYLOAD_LIMIT = 256 * 1_024;
 const TYPING_TIMEOUT_MS = 5_000;
 const RATE_LIMIT_WINDOW_MS = 15 * 60 * 1_000;
@@ -101,117 +102,6 @@ const requireModule = createRequire(import.meta.url);
 const webSockets: typeof Ws = requireModule(join(dirname(requireModule.resolve("ws/package.json")), "index.js"));
 
 const logger = createOpenBotLogger("team-api-server");
-
-type TeamApiAgentMethods = Pick<
-  AgentService,
-  | "getStatus"
-  | "getRuntimeSnapshot"
-  | "getUsage"
-  | "listModels"
-  | "listBots"
-  | "listConversationReads"
-  | "createBot"
-  | "committedBotDuplication"
-  | "duplicateBot"
-  | "commitBotDuplication"
-  | "updateBot"
-  | "deleteBot"
-  | "listMemories"
-  | "createMemory"
-  | "updateMemory"
-  | "deleteMemory"
-  | "clearMemories"
-  | "listRoutines"
-  | "createRoutine"
-  | "updateRoutine"
-  | "deleteRoutine"
-  | "testRoutine"
-  | "listRoutineRuns"
-  | "setAvatar"
-  | "resolveAvatar"
-  | "readConversationFor"
-  | "readConversationPageFor"
-  | "searchConversationMessages"
-  | "markConversationRead"
-  | "markConversationUnread"
-  | "prepareImportedAttachments"
-  | "discardDraftAttachment"
-  | "resolveSharedFile"
-  | "resolveWorkspaceFile"
-  | "sendMessage"
-  | "listQueue"
-  | "acknowledgeFailedTurn"
-  | "setMessageReaction"
-  | "cancelQueuedMessage"
-  | "steerQueuedMessage"
-  | "updateQueuedMessage"
-  | "reorderQueue"
-  | "interrupt"
-  | "respondToPrompt"
-  | "respondToApproval"
-  | "respondToBrowserTakeover"
->;
-
-type TeamApiAgents = TeamApiAgentMethods & {
-  on: (event: "event", listener: (event: AgentEvent) => void) => void;
-  off: (event: "event", listener: (event: AgentEvent) => void) => void;
-};
-
-type TeamApiMailbox = Pick<MailboxStore, "resolveAttachment">;
-type TeamApiSidebarLayout = Pick<
-  SidebarLayoutStore,
-  "getSnapshot" | "mutate" | "removeAgent" | "placeDuplicateAfter"
-> & {
-  on: (event: "changed", listener: (layout: SidebarLayoutSnapshot) => void) => void;
-  off: (event: "changed", listener: (layout: SidebarLayoutSnapshot) => void) => void;
-};
-type TeamApiBrowser = Pick<
-  BrowserHost,
-  | "listTabs"
-  | "getControlState"
-  | "open"
-  | "activate"
-  | "navigate"
-  | "reload"
-  | "close"
-  | "capturePreview"
-  | "setVisible"
->;
-type TeamApiRemoteScreen = Pick<
-  RemoteScreenGateway,
-  | "handlesUpgrade"
-  | "handleUpgrade"
-  | "handlesHttp"
-  | "handleHttp"
-  | "stop"
-  | "capabilities"
-  | "createSession"
-  | "selectDisplay"
-  | "closeMemberSession"
-  | "revokeTeamSession"
-  | "revokeMember"
->;
-
-interface TeamApiOptions {
-  appVersion?: string;
-  store: TeamStore;
-  agents: TeamApiAgents;
-  skills?: { listInstalledForChatTags: (botId: string) => Promise<InstalledSkill[]> };
-  sidebarLayout?: TeamApiSidebarLayout;
-  mailbox: TeamApiMailbox;
-  browser: TeamApiBrowser;
-  remoteScreen?: TeamApiRemoteScreen;
-  redeemCentralTicket?: (ticket: string, serverId: string) => Promise<CentralAuthUser | null>;
-  onPresence?: (snapshot: TeamPresenceSnapshot) => void;
-  chat?: TeamChatStore;
-  onDirectMessage?: (event: DirectMessageRealtimeEvent) => void;
-  onDirectTyping?: (event: DirectTypingRealtimeEvent) => void;
-  createInvite?: (input: CreateTeamInviteInput) => Promise<InviteSummary>;
-  onSessionRevoked?: (sessionId: string) => Promise<void> | void;
-  rateLimitCapacity?: number;
-  now?: () => number;
-  logger?: Logger;
-}
 
 interface EventClientState {
   token: string;
@@ -1770,82 +1660,6 @@ function unavailableSidebarLayout(): TeamApiSidebarLayout {
   };
 }
 
-function requestCapabilities(request: import("node:http").IncomingMessage): Set<string> {
-  const header = request.headers[TEAM_CAPABILITIES_HEADER.toLowerCase()];
-  const value = Array.isArray(header) ? header.join(",") : header;
-  if (!value || value.length > 4_096) return new Set();
-  const capabilities = value.split(",").map((capability) => capability.trim());
-  if (capabilities.length > 64) return new Set();
-  return new Set(capabilities.filter(isTeamCurrentCapability));
-}
-
-function conversationSnapshotForCapabilities(
-  snapshot: ConversationSnapshot,
-  capabilities: ReadonlySet<string>,
-): ConversationSnapshot {
-  return {
-    ...snapshot,
-    messages: snapshot.messages.filter((message) => markerSupported(message.itemType, capabilities)),
-  };
-}
-
-function conversationForCapabilities(
-  conversation: ConversationWithReadState,
-  capabilities: ReadonlySet<string>,
-): ConversationWithReadState {
-  const messages = conversation.messages.filter((message) => markerSupported(message.itemType, capabilities));
-  if (!conversation.readState) return { ...conversation, messages };
-  return {
-    ...conversation,
-    messages,
-    readState: {
-      ...conversation.readState,
-      throughMessageId: supportedConversationCursor(
-        conversation.messages,
-        conversation.readState.throughMessageId,
-        capabilities,
-      ),
-    },
-  };
-}
-
-function supportedConversationCursor(
-  messages: ConversationSnapshot["messages"],
-  throughMessageId: string | null,
-  capabilities: ReadonlySet<string>,
-): string | null {
-  if (!throughMessageId) return null;
-  const boundary = messages.findIndex((message) => message.id === throughMessageId);
-  for (let index = boundary; index >= 0; index -= 1) {
-    const message = messages[index];
-    if (message && markerSupported(message.itemType, capabilities)) return message.id;
-  }
-  return null;
-}
-
-function markerExclusionsForCapabilities(capabilities: ReadonlySet<string>): {
-  excludeRoutineEvents: boolean;
-  excludeRoutineRunEvents: boolean;
-  excludeHostedSiteEvents: boolean;
-} {
-  return {
-    excludeRoutineEvents: !capabilities.has("routine-event-markers"),
-    excludeRoutineRunEvents: !capabilities.has("routine-run-event-markers"),
-    excludeHostedSiteEvents: !capabilities.has("hosted-site-event-markers"),
-  };
-}
-
-function markerSupported(itemType: string | undefined, capabilities: ReadonlySet<string>): boolean {
-  if (itemType?.startsWith(ROUTINE_EVENT_ITEM_TYPE_PREFIX)) return capabilities.has("routine-event-markers");
-  if (itemType?.startsWith(ROUTINE_RUN_EVENT_ITEM_TYPE_PREFIX)) {
-    return capabilities.has("routine-run-event-markers");
-  }
-  if (itemType?.startsWith(HOSTED_SITE_EVENT_ITEM_TYPE_PREFIX)) {
-    return capabilities.has("hosted-site-event-markers");
-  }
-  return true;
-}
-
 function eventCapability(event: AgentEvent): TeamCurrentCapability | null {
   if (event.type === "turn-progress") return TEAM_AGENT_ACTIVITY_CAPABILITY;
   if (event.type === "runtime-snapshot") return "agent-runtime-snapshots";
@@ -1853,295 +1667,4 @@ function eventCapability(event: AgentEvent): TeamCurrentCapability | null {
   if (event.type === "browser-changed" || event.type === "browser-control-changed") return "browser-control";
   if (event.type === "conversation-page") return "conversation-pagination";
   return null;
-}
-
-class HttpError extends Error {
-  constructor(
-    readonly status: number,
-    message: string,
-  ) {
-    super(message);
-  }
-}
-
-function bearerToken(value: string | undefined): string | null {
-  const match = value?.match(/^Bearer ([A-Za-z0-9_-]{20,512})$/);
-  return match?.[1] ?? null;
-}
-
-function publicHttpBaseUrl(request: import("node:http").IncomingMessage): string {
-  const forwardedHost = firstHeaderValue(request.headers["x-forwarded-host"]);
-  const host = forwardedHost || request.headers.host;
-  if (!host || !/^[A-Za-z0-9.:[\]-]+$/.test(host)) {
-    throw new HttpError(400, "A valid public host is required.");
-  }
-  const forwardedProtocol = firstHeaderValue(request.headers["x-forwarded-proto"]);
-  const protocol = forwardedProtocol === "https" ? "https" : "http";
-  return `${protocol}://${host}`;
-}
-
-function firstHeaderValue(value: string | string[] | undefined): string | undefined {
-  const first = Array.isArray(value) ? value[0] : value?.split(",", 1)[0];
-  return first?.trim();
-}
-
-function requireAdmin(member: TeamMemberSummary): void {
-  if (member.role === "member") throw new HttpError(403, "Administrator access is required.");
-}
-
-function parseBrowserBounds(value: unknown): {
-  x: number;
-  y: number;
-  width: number;
-  height: number;
-} {
-  if (!isDynamicRecord(value)) throw new HttpError(400, "Invalid browser bounds.");
-  const x = value.x;
-  const y = value.y;
-  const width = value.width;
-  const height = value.height;
-  if (
-    !isNumber(x) ||
-    !isNumber(y) ||
-    !isNumber(width) ||
-    !isNumber(height) ||
-    ![x, y, width, height].every(Number.isFinite)
-  ) {
-    throw new HttpError(400, "Invalid browser bounds.");
-  }
-  return { x, y, width, height };
-}
-
-async function readJson(request: import("node:http").IncomingMessage): Promise<DynamicRecord> {
-  const chunks: Buffer[] = [];
-  let size = 0;
-  for await (const chunk of request) {
-    const bytes = Buffer.isBuffer(chunk) ? chunk : Buffer.from(chunk);
-    size += bytes.length;
-    if (size > JSON_LIMIT) throw new HttpError(413, "Request body is too large.");
-    chunks.push(bytes);
-  }
-  try {
-    const value = JSON.parse(Buffer.concat(chunks).toString("utf8"));
-    return requestProtocol(request) === TEAM_PROTOCOL_V3
-      ? decodeTeamProtocolV3CurrentHttpRequest(request.method ?? "GET", request.url ?? "/", value, {
-          preserveSemanticTags: supportsTeamSemanticTags(requestCapabilities(request)),
-        })
-      : decodeTeamProtocolV1CurrentHttpRequest(request.method ?? "GET", request.url ?? "/", value);
-  } catch {
-    throw new HttpError(400, "A valid JSON object is required.");
-  }
-}
-
-function requestProtocol(request: import("node:http").IncomingMessage): number {
-  const raw = firstHeaderValue(request.headers[TEAM_PROTOCOL_VERSION_HEADER.toLowerCase()]);
-  const protocol = raw ? Number(raw) : TEAM_PROTOCOL_V1;
-  return Number.isSafeInteger(protocol) ? protocol : TEAM_PROTOCOL_V1;
-}
-
-function stringField(
-  value: DynamicRecord,
-  field: string,
-  allowEmpty = false,
-  maxLength: number = INPUT_LIMITS.identifier,
-): string {
-  const item = value[field];
-  if (!isString(item) || (!allowEmpty && !item.trim())) {
-    throw new HttpError(400, `${field} is required.`);
-  }
-  if (item.length > maxLength) throw new HttpError(400, `${field} is too long.`);
-  return item;
-}
-
-function nullableString(
-  value: DynamicRecord,
-  field: string,
-  maxLength: number = INPUT_LIMITS.identifier,
-): string | null {
-  const item = value[field];
-  if (item === undefined || item === null || item === "") return null;
-  if (!isString(item)) throw new HttpError(400, `${field} must be a string.`);
-  if (item.length > maxLength) throw new HttpError(400, `${field} is too long.`);
-  return item;
-}
-
-function pathIdentifier(value: string | undefined, field: string): string {
-  let decoded: string;
-  try {
-    decoded = decodeURIComponent(value ?? "");
-  } catch {
-    throw new HttpError(400, `${field} is invalid.`);
-  }
-  if (!decoded || decoded.length > INPUT_LIMITS.identifier) {
-    throw new HttpError(400, `${field} is invalid.`);
-  }
-  return decoded;
-}
-
-function stringArray(
-  value: DynamicRecord,
-  field: string,
-  maxItems: number = INPUT_LIMITS.attachments,
-  maxLength: number = INPUT_LIMITS.identifier,
-): string[] {
-  const item = value[field];
-  if (item === undefined) return [];
-  if (
-    !Array.isArray(item) ||
-    item.length > maxItems ||
-    !item.every((entry) => isString(entry) && entry.length <= maxLength)
-  ) {
-    throw new HttpError(400, `${field} must be a string array.`);
-  }
-  return item;
-}
-
-function promptRequestId(value: unknown): string | number {
-  if (isNumber(value) && Number.isSafeInteger(value)) return value;
-  if (isString(value) && value.length > 0 && value.length <= INPUT_LIMITS.identifier) {
-    return value;
-  }
-  throw new HttpError(400, "requestId is invalid.");
-}
-
-function promptAnswers(value: unknown): Record<string, string[]> {
-  if (!isDynamicRecord(value)) {
-    throw new HttpError(400, "answers is required.");
-  }
-  const entries = Object.entries(value);
-  if (entries.length > INPUT_LIMITS.promptQuestions) {
-    throw new HttpError(400, "Too many prompt answers.");
-  }
-  const answers: Record<string, string[]> = {};
-  let totalTextLength = 0;
-  for (const [key, answer] of entries) {
-    if (
-      key.length > INPUT_LIMITS.identifier ||
-      !Array.isArray(answer) ||
-      answer.length > INPUT_LIMITS.promptAnswersPerQuestion ||
-      !answer.every((item) => isString(item) && item.length <= INPUT_LIMITS.promptAnswerText)
-    ) {
-      throw new HttpError(400, "A prompt answer is invalid.");
-    }
-    totalTextLength += answer.reduce((length, item) => length + item.length, 0);
-    if (totalTextLength > INPUT_LIMITS.promptAnswersTotalText) {
-      throw new HttpError(400, "Prompt answers are too long.");
-    }
-    answers[key] = answer;
-  }
-  return answers;
-}
-
-function approvalDecision(value: unknown): RespondToApprovalInput["decision"] {
-  if (value === "accept" || value === "decline") return value;
-  throw new HttpError(400, "approval decision is invalid.");
-}
-
-function browserTakeoverDecision(value: unknown): RespondToBrowserTakeoverInput["decision"] {
-  if (value === "complete" || value === "cancel") return value;
-  throw new HttpError(400, "browser takeover decision is invalid.");
-}
-
-function botUpdate(value: DynamicRecord, botId: string): UpdateBotInput {
-  if (value.role !== undefined) throw new HttpError(400, "role is invalid.");
-  const result: UpdateBotInput = { botId };
-  const textFields = {
-    name: INPUT_LIMITS.agentName,
-    title: INPUT_LIMITS.agentTitle,
-    description: INPUT_LIMITS.agentDescription,
-  } as const;
-  for (const [field, maxLength] of Object.entries(textFields)) {
-    const item = value[field];
-    if (item === undefined) continue;
-    if (!isString(item) || item.length > maxLength) {
-      throw new HttpError(400, `${field} is invalid.`);
-    }
-    if (field === "name") result.name = item;
-    else if (field === "title") result.title = item;
-    else result.description = item;
-  }
-  if (value.notifications !== undefined) {
-    if (!isBoolean(value.notifications)) {
-      throw new HttpError(400, "notifications is invalid.");
-    }
-    result.notifications = value.notifications;
-  }
-  if (value.provider !== undefined) {
-    if (value.provider !== "codex" && value.provider !== "claude" && value.provider !== "grok") {
-      throw new HttpError(400, "provider is invalid.");
-    }
-    result.provider = value.provider;
-  }
-  if (value.model !== undefined) {
-    if (!isAgentModel(value.model)) throw new HttpError(400, "model is invalid.");
-    result.model = value.model;
-  }
-  if (value.reasoningEffort !== undefined) {
-    if (!isReasoningEffort(value.reasoningEffort)) {
-      throw new HttpError(400, "reasoningEffort is invalid.");
-    }
-    result.reasoningEffort = value.reasoningEffort;
-  }
-  if (value.avatarSeed !== undefined) {
-    if (!isAvatarSeed(value.avatarSeed)) throw new HttpError(400, "avatarSeed is invalid.");
-    result.avatarSeed = value.avatarSeed;
-  }
-  if (value.avatarHue !== undefined) {
-    if (value.avatarHue !== null && !isAvatarHue(value.avatarHue)) {
-      throw new HttpError(400, "avatarHue is invalid.");
-    }
-    result.avatarHue = value.avatarHue;
-  }
-  return result;
-}
-
-function botCreate(value: DynamicRecord): CreateBotInput {
-  const avatarHue = value.avatarHue;
-  if (!isAvatarSeed(value.avatarSeed)) throw new HttpError(400, "avatarSeed is invalid.");
-  if (avatarHue !== null && !isAvatarHue(avatarHue)) throw new HttpError(400, "avatarHue is invalid.");
-  return {
-    name: requiredCreateText(value.name, "name", INPUT_LIMITS.agentName),
-    description: requiredCreateText(value.description, "description", INPUT_LIMITS.agentDescription),
-    avatarSeed: value.avatarSeed,
-    avatarHue,
-    initialMessage: requiredCreateText(value.initialMessage, "initialMessage", INPUT_LIMITS.messageText),
-  };
-}
-
-function requiredCreateText(value: unknown, field: string, maximum: number): string {
-  if (!isString(value) || !value.trim() || value.length > maximum) {
-    throw new HttpError(400, `${field} is invalid.`);
-  }
-  return value;
-}
-
-async function readBinary(request: import("node:http").IncomingMessage, limit: number): Promise<Uint8Array> {
-  const chunks: Buffer[] = [];
-  let size = 0;
-  for await (const chunk of request) {
-    const bytes = Buffer.isBuffer(chunk) ? chunk : Buffer.from(chunk);
-    size += bytes.length;
-    if (size > limit) throw new HttpError(413, "Attachment exceeds the 100 MB limit.");
-    chunks.push(bytes);
-  }
-  return new Uint8Array(Buffer.concat(chunks));
-}
-
-function pageAnchor(url: URL): ConversationPageAnchor {
-  const before = url.searchParams.get("before");
-  const around = url.searchParams.get("around");
-  if (before && around) throw new HttpError(400, "Choose one conversation page anchor.");
-  if (before) return { type: "before", cursor: before };
-  if (around) return { type: "around", messageId: around };
-  return { type: "latest" };
-}
-
-function pageLimit(url: URL): number {
-  const raw = url.searchParams.get("limit");
-  if (raw === null) return 50;
-  const value = Number(raw);
-  if (!Number.isInteger(value) || value < 1 || value > 100) {
-    throw new HttpError(400, "The conversation page limit must be between 1 and 100.");
-  }
-  return value;
 }
